Fix bug
In some situation a wrapper was unable to find a constant String object.
This was due to the fact that the pre-frontier runner added a String
object introduced in the post-frontier state generated by the runner.
When exploring other post-frontier states corresponding to the same
pre-frontier, the String object may not exist. This fix corrects the
pre-frontier runner so it does not adds the String objects in the
post-frontier state it reaches.