Fix initial sync of TerminationManager, refactor PerformerJBSE
With the bootstrap phase that in some cases is not done by seeding, it
is possible that the system appears idle to the termination manager
(empty queues, no task to the thread pools) but it is actually
bootstrapping. Now the termination manager starts detecting inactivity
of the performers only after it has detected some activity (timeout
management is unchanged).