Fix Performer's code to better match with indices
Eliminated prefetching of jobs by performer from input queue: Now a job
is fetched only when it is dispatched to a server that elaborates it.
Similarly, the PerformerEvosuiteRMI does not block any longer on an
Evosuite server without availability, but always dispatches to the best
server it can find (users can avoid that servers are overloaded beyond
their availability by exploiting the throttling parameter).